move es client constructor to es app
