+++ backend/indexing/models.py
@@ -1,7 +1,7 @@
 from django.db import models
 from elasticsearch import Elasticsearch
 
-from ianalyzer.elasticsearch import elasticsearch
+from es.client import elasticsearch
 from addcorpus.models import Corpus
 from es.models import Index
